# doc/api.yaml
{"openapi":"3.0.0","paths":{"/user":{"get":{"operationId":"UsersController_getUsers","parameters":[],"responses":{"200":{"description":""}}},"post":{"operationId":"UsersController_createUser","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateUserDto"}}}},"responses":{"201":{"description":""}}}},"/user/{id}":{"get":{"operationId":"UsersController_getUser","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}}},"put":{"operationId":"UsersController_updateUser","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateUserDto"}}}},"responses":{"200":{"description":""}}},"delete":{"operationId":"UsersController_deleteUser","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"204":{"description":""}}}},"/track":{"get":{"operationId":"TracksController_getTracks","parameters":[],"responses":{"200":{"description":""}}},"post":{"operationId":"TracksController_createTrack","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateTrackDto"}}}},"responses":{"201":{"description":""}}}},"/track/{id}":{"get":{"operationId":"TracksController_getTrack","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}}},"put":{"operationId":"TracksController_updateTrack","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateTrackDto"}}}},"responses":{"200":{"description":""}}},"delete":{"operationId":"TracksController_deleteTrack","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"204":{"description":""}}}},"/favs":{"get":{"operationId":"FavouritesController_getAllFavourites","parameters":[],"responses":{"200":{"description":""}}}},"/favs/track/{id}":{"post":{"operationId":"FavouritesController_addTrack","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"201":{"description":""}}},"delete":{"operationId":"FavouritesController_deleteTrack","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"204":{"description":""}}}},"/favs/album/{id}":{"post":{"operationId":"FavouritesController_addAlbum","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"201":{"description":""}}},"delete":{"operationId":"FavouritesController_deleteAlbum","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"204":{"description":""}}}},"/favs/artist/{id}":{"post":{"operationId":"FavouritesController_addArtist","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"201":{"description":""}}},"delete":{"operationId":"FavouritesController_deleteArtist","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"204":{"description":""}}}},"/album":{"get":{"operationId":"AlbumsController_getArtists","parameters":[],"responses":{"200":{"description":""}}},"post":{"operationId":"AlbumsController_createAlbum","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateAlbumDto"}}}},"responses":{"201":{"description":""}}}},"/album/{id}":{"get":{"operationId":"AlbumsController_getAlbum","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}}},"put":{"operationId":"AlbumsController_updateAlbum","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateAlbumDto"}}}},"responses":{"200":{"description":""}}},"delete":{"operationId":"AlbumsController_deleteAlbum","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"204":{"description":""}}}},"/artist":{"get":{"operationId":"ArtistsController_getArtists","parameters":[],"responses":{"200":{"description":""}}},"post":{"operationId":"ArtistsController_createArtist","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateArtistDto"}}}},"responses":{"201":{"description":""}}}},"/artist/{id}":{"get":{"operationId":"ArtistsController_getArtist","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}}},"put":{"operationId":"ArtistsController_updateArtist","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateArtistDto"}}}},"responses":{"200":{"description":""}}},"delete":{"operationId":"ArtistsController_deleteArtist","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"204":{"description":""}}}}},"info":{"title":"Home Library Service","description":"Home music library service","version":"1.0","contact":{}},"tags":[],"servers":[],"components":{"schemas":{"CreateUserDto":{"type":"object","properties":{}},"UpdateUserDto":{"type":"object","properties":{}},"CreateTrackDto":{"type":"object","properties":{}},"UpdateTrackDto":{"type":"object","properties":{}},"CreateAlbumDto":{"type":"object","properties":{}},"UpdateAlbumDto":{"type":"object","properties":{}},"CreateArtistDto":{"type":"object","properties":{}},"UpdateArtistDto":{"type":"object","properties":{}}}}}